requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# OpenAI
openai==1.10.0

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Vector DB
chromadb==0.4.22

# Scientific computing (for embeddings)
numpy==1.26.0

# Telegram
python-telegram-bot==20.7

# Monitoring
prometheus-client==0.19.0

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utils
feedparser==6.0.10
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.8.2
tenacity==8.2.3

# Logging
structlog==24.1.0

